+++ pkg/model/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "bootstrapscript.go",
         "context.go",
+        "issuerdiscovery.go",
         "manifests.go",
         "master_volumes.go",
         "names.go",
@@ -47,6 +48,7 @@ go_library(
         "//util/pkg/vfs:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2:go_default_library",
         "//vendor/github.com/blang/semver/v4:go_default_library",
+        "//vendor/gopkg.in/square/go-jose.v2:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/net:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/yaml:go_default_library",
