+++ pkg/model/awsmodel/oidc_provider.go
@@ -17,21 +17,10 @@ limitations under the License.
 package awsmodel
 
 import (
-	"bytes"
-	"crypto"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/json"
-	"encoding/pem"
-	"fmt"
-	"io"
-
-	"gopkg.in/square/go-jose.v2"
 	"k8s.io/kops/pkg/featureflag"
 	"k8s.io/kops/pkg/model/iam"
 	"k8s.io/kops/upup/pkg/fi"
 	"k8s.io/kops/upup/pkg/fi/cloudup/awstasks"
-	"k8s.io/kops/upup/pkg/fi/fitasks"
 )
 
 // OIDCProviderBuilder configures IAM OIDC Provider
@@ -39,16 +28,6 @@ type OIDCProviderBuilder struct {
 	*AWSModelContext
 	KeyStore  fi.CAStore
 	Lifecycle *fi.Lifecycle
-}
-
-type oidcDiscovery struct {
-	Issuer                string   `json:"issuer"`
-	JWKSURI               string   `json:"jwks_uri"`
-	AuthorizationEndpoint string   `json:"authorization_endpoint"`
-	ResponseTypes         []string `json:"response_types_supported"`
-	SubjectTypes          []string `json:"subject_types_supported"`
-	SigningAlgs           []string `json:"id_token_signing_alg_values_supported"`
-	ClaimsSupported       []string `json:"claims_supported"`
 }
 
 var _ fi.ModelBuilder = &OIDCProviderBuilder{}
@@ -68,11 +47,6 @@ func (b *OIDCProviderBuilder) Build(c *fi.ModelBuilderContext) error {
 		return err
 	}
 
-	signingKeyTaskObject, found := c.Tasks["Keypair/service-account"]
-	if !found {
-		return fmt.Errorf("keypair/service-account task not found")
-	}
-
 	fingerprints := getFingerprints()
 
 	thumbprints := []*string{}
@@ -80,36 +54,6 @@ func (b *OIDCProviderBuilder) Build(c *fi.ModelBuilderContext) error {
 	for _, fingerprint := range fingerprints {
 		thumbprints = append(thumbprints, fi.String(fingerprint))
 	}
-
-	skTask := signingKeyTaskObject.(*fitasks.Keypair)
-
-	keys := &OIDCKeys{
-		SigningKey: skTask,
-	}
-
-	discovery, err := buildDiscoveryJSON(serviceAccountIssuer)
-	if err != nil {
-		return err
-	}
-	keysFile := &fitasks.ManagedFile{
-		Contents:  keys,
-		Lifecycle: b.Lifecycle,
-		Location:  fi.String("oidc/keys.json"),
-		Name:      fi.String("keys.json"),
-		Base:      fi.String(b.Cluster.Spec.PublicDataStore),
-		Public:    fi.Bool(true),
-	}
-	c.AddTask(keysFile)
-
-	discoveryFile := &fitasks.ManagedFile{
-		Contents:  fi.NewBytesResource(discovery),
-		Lifecycle: b.Lifecycle,
-		Location:  fi.String("oidc/.well-known/openid-configuration"),
-		Name:      fi.String("discovery.json"),
-		Base:      fi.String(b.Cluster.Spec.PublicDataStore),
-		Public:    fi.Bool(true),
-	}
-	c.AddTask(discoveryFile)
 
 	c.AddTask(&awstasks.IAMOIDCProvider{
 		Name:        fi.String(b.ClusterName()),
@@ -121,76 +65,6 @@ func (b *OIDCProviderBuilder) Build(c *fi.ModelBuilderContext) error {
 	})
 
 	return nil
-}
-
-func buildDiscoveryJSON(issuerURL string) ([]byte, error) {
-	d := oidcDiscovery{
-		Issuer:                fmt.Sprintf("%v/", issuerURL),
-		JWKSURI:               fmt.Sprintf("%v/keys.json", issuerURL),
-		AuthorizationEndpoint: "urn:kubernetes:programmatic_authorization",
-		ResponseTypes:         []string{"id_token"},
-		SubjectTypes:          []string{"public"},
-		SigningAlgs:           []string{"RS256"},
-		ClaimsSupported:       []string{"sub", "iss"},
-	}
-	return json.MarshalIndent(d, "", "")
-}
-
-type KeyResponse struct {
-	Keys []jose.JSONWebKey `json:"keys"`
-}
-
-type OIDCKeys struct {
-	SigningKey *fitasks.Keypair
-}
-
-// GetDependencies adds CA to the list of dependencies
-func (o *OIDCKeys) GetDependencies(tasks map[string]fi.Task) []fi.Task {
-	return []fi.Task{
-		o.SigningKey,
-	}
-}
-func (o *OIDCKeys) Open() (io.Reader, error) {
-
-	certBytes, err := fi.ResourceAsBytes(o.SigningKey.Certificate())
-	if err != nil {
-		return nil, fmt.Errorf("failed to get cert: %w", err)
-	}
-	block, _ := pem.Decode(certBytes)
-	cert, err := x509.ParseCertificate(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse cert: %w", err)
-	}
-
-	publicKey := cert.PublicKey
-
-	publicKeyDERBytes, err := x509.MarshalPKIXPublicKey(publicKey)
-	if err != nil {
-		return nil, fmt.Errorf("failed to serialize public key to DER format: %v", err)
-	}
-
-	hasher := crypto.SHA256.New()
-	hasher.Write(publicKeyDERBytes)
-	publicKeyDERHash := hasher.Sum(nil)
-
-	keyID := base64.RawURLEncoding.EncodeToString(publicKeyDERHash)
-
-	keys := []jose.JSONWebKey{
-		{
-			Key:       publicKey,
-			KeyID:     keyID,
-			Algorithm: string(jose.RS256),
-			Use:       "sig",
-		},
-	}
-
-	keyResponse := KeyResponse{Keys: keys}
-	jsonBytes, err := json.MarshalIndent(keyResponse, "", "")
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal json: %w", err)
-	}
-
-	return bytes.NewReader(jsonBytes), nil
 }
 
 func getFingerprints() []string {
